Merged PR 66689: Create and Clone visual - add flag to control whether we should focus on the added visual
**Update**
It has been decided to change the implementation to add a flag to each function (createVisual and clone) called autoFocus, I will update it later.

Right now, when adding a visual in the portal or in embedded u
Right now, when adding a visual in the portal or in embedded using createVisual and clone the page moves to focus on the added visual.
We want to add a flag to the settings to control this behavior.

**Note fore reviewers:**
The name is not yet finalized, feel free to add suggestions.